+++ src/Simplify.Web/StaticFiles/Handlers/ClientCachedFileHandler.cs
@@ -1,16 +1,17 @@
+using System.Net;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
-using Simplify.Web.Http.ResponseWriting;
+using Simplify.Web.Http.Mime;
+using Simplify.Web.Http.ResponseTime;
 using Simplify.Web.StaticFiles.Context;
-using Simplify.Web.StaticFiles.IO;
 
 namespace Simplify.Web.StaticFiles.Handlers;
 
 /// <summary>
-/// Provides the new file handler.
+/// Provides the cached on client side files handler.
 /// </summary>
 /// <seealso cref="IStaticFileRequestHandler" />
-public class NewFileHandler(IResponseWriter responseWriter, IStaticFile staticFile) : IStaticFileRequestHandler
+public class ClientCachedFileHandler : IStaticFileRequestHandler
 {
 	/// <summary>
 	/// Determines whether this handler can handle the file requested.
@@ -19,17 +20,19 @@ public class NewFileHandler(IResponseWriter responseWriter, IStaticFile staticFi
 	/// <returns>
 	///   <c>true</c> if this handler can handle the file requested; otherwise, <c>false</c>.
 	/// </returns>
-	public bool CanHandle(IStaticFileProcessingContext context) => !context.CanBeCached;
+	public bool CanHandle(IStaticFileProcessingContext context) => context.CanBeCached;
 
 	/// <summary>
 	/// Executes the handler asynchronously.
 	/// </summary>
 	/// <param name="context">The context.</param>
 	/// <param name="response">The response.</param>
-	public async Task ExecuteAsync(IStaticFileProcessingContext context, HttpResponse response)
+	public Task ExecuteAsync(IStaticFileProcessingContext context, HttpResponse response)
 	{
-		response.SetNewReturningFileAttributes(context);
+		response.SetContentMimeType(context.RelativeFilePath);
+		response.SetLastModifiedTime(context.LastModificationTime);
+		response.StatusCode = (int)HttpStatusCode.NotModified;
 
-		await responseWriter.WriteAsync(response, await staticFile.GetDataAsync(context.RelativeFilePath));
+		return Task.CompletedTask;
 	}
 }