+++ src/Simplify.Web/StaticFiles/Handlers/InMemoryFilesCacheHandler.cs
@@ -1,17 +1,17 @@
-using System.Net;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
-using Simplify.Web.Http.Mime;
-using Simplify.Web.Http.ResponseTime;
+using Simplify.Web.Http.ResponseWriting;
+using Simplify.Web.StaticFiles.Cache;
 using Simplify.Web.StaticFiles.Context;
+using Simplify.Web.StaticFiles.IO;
 
 namespace Simplify.Web.StaticFiles.Handlers;
 
 /// <summary>
-/// Provides the cached file handler.
+/// Provides the in-memory files cache.
 /// </summary>
 /// <seealso cref="IStaticFileRequestHandler" />
-public class CachedFileHandler : IStaticFileRequestHandler
+public class InMemoryFilesCacheHandler(IResponseWriter responseWriter, IStaticFile staticFile) : IStaticFileRequestHandler
 {
 	/// <summary>
 	/// Determines whether this handler can handle the file requested.
@@ -20,19 +20,17 @@ public class CachedFileHandler : IStaticFileRequestHandler
 	/// <returns>
 	///   <c>true</c> if this handler can handle the file requested; otherwise, <c>false</c>.
 	/// </returns>
-	public bool CanHandle(IStaticFileProcessingContext context) => context.CanBeCached;
+	public bool CanHandle(IStaticFileProcessingContext context) => !context.CanBeCached;
 
 	/// <summary>
 	/// Executes the handler asynchronously.
 	/// </summary>
 	/// <param name="context">The context.</param>
 	/// <param name="response">The response.</param>
-	public Task ExecuteAsync(IStaticFileProcessingContext context, HttpResponse response)
+	public async Task ExecuteAsync(IStaticFileProcessingContext context, HttpResponse response)
 	{
-		response.SetContentMimeType(context.RelativeFilePath);
-		response.SetLastModifiedTime(context.LastModificationTime);
-		response.StatusCode = (int)HttpStatusCode.NotModified;
+		response.SetNewReturningFileAttributes(context);
 
-		return Task.CompletedTask;
+		await responseWriter.WriteAsync(response, FilesInMemoryCache.Items.GetOrAdd(context.RelativeFilePath, staticFile.GetData));
 	}
 }